+++ app/controllers/proficient_projects_controller.rb
@@ -259,12 +259,16 @@ class ProficientProjectsController < DevelopmentProgramsController
   end
 
   def update_videos
-    if params['deletevideos'].present?
+    videos_id = params['deletevideos']
+    if videos_id.present?
+      videos_id = videos_id.split(',').uniq.map{|id| id.to_i}
       @proficient_project.videos.each do |f|
-        if (params['deletevideos'] - f.video.map{|v| v.filename.to_s}).any?
-          # Code more here
-          f.video.purge
-          f.destroy
+        if (f.video.pluck(:id) & videos_id).any?
+          videos_id.each do |video_id|
+            video = f.video.find(video_id)
+            video.purge
+          end
+          f.destroy unless f.video.attached?
         end
       end
     end
